refactor(spf): move dom/network to top-level src/network
fetch, Request, Response, and ReadableStream are universal JavaScript
APIs — available in browsers, Node 18+, Deno, and workers. They live
under src/dom/ only for historical reasons. Moving them out:

- Breaks a layering violation surfaced by the core tsconfig enforcement:
  src/media/behaviors/resolve-presentation.ts and resolve-track.ts
  reached into src/dom/network/fetch, which was a media → dom import
  that couldn't be modeled cleanly with project references (dom already
  depends on media, adding the reverse edge creates a cycle).

- Gives a natural home for future transport work. When DASH and MoQ
  introduce WebTransport or custom transports, they can land as
  siblings under src/network/ without reshaping dom/.

Changes:
- src/dom/network/{fetch,chunked-stream-iterable}.ts → src/network/
- tests move alongside
- packages/spf/vitest.config.ts: add a network project (Node env;
  these tests don't need a browser)
- import paths updated in all.ts, dom/behaviors/load-segments.ts,
  media/behaviors/resolve-{presentation,track}.ts, and the fetch.ts
  file's own media/types import

No runtime change. Build, tests (707), typecheck, lint, workspace
check all green.
