+++ packages/spf/src/dom/behaviors/load-segments.ts
@@ -5,6 +5,8 @@ import { DEFAULT_FORWARD_BUFFER_CONFIG } from '../../media/buffer/forward-buffer
 import type { AddressableObject, Presentation, ResolvedTrack } from '../../media/types';
 import { isResolvedTrack } from '../../media/types';
 import { getSelectedTrack, type TrackSelectionState } from '../../media/utils/track-selection';
+import { ChunkedStreamIterable, type ChunkedStreamIterableOptions } from '../../network/chunked-stream-iterable';
+import { fetchResolvable } from '../../network/fetch';
 import {
   type BufferState,
   createSegmentLoaderActor,
@@ -12,8 +14,6 @@ import {
   type SourceBufferState,
 } from '../actors/segment-loader';
 import type { SourceBufferActor } from '../media/source-buffer-actor';
-import { ChunkedStreamIterable, type ChunkedStreamIterableOptions } from '../network/chunked-stream-iterable';
-import { fetchResolvable } from '../network/fetch';
 import type { MediaTrackType } from './setup-sourcebuffer';
 
 // Re-export buffer state types for consumers that import them from this module.
